switched to named code blocks for easier slide management
