+++ presentations/async-await-intro/fetch_into_string.rs
